requirements.txt
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.4
boto3==1.16.53
botocore==1.19.53
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
distlib==0.3.1
filelock==3.0.12
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pdfminer.six==20201018
Pillow==8.1.0
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
PyMuPDF==1.18.6
python-dateutil==2.8.1
s3transfer==0.3.4
six==1.15.0
sortedcontainers==2.3.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
